Change Interrupt API
Return whether there was an interruption. It's kind of logical you want to know if the inner block is succeeded. It can always be done by adding supplementary variables but it always looked as some cumbersome C-style to me. I stumbled upon this again when I tried to explain the effect in docs.